+++ gazebo/gui/building/BuildingEditor.cc
@@ -45,23 +45,23 @@ BuildingEditor::BuildingEditor(MainWindow *_mainWindow)
   this->Init("buildingEditorTab", "Building Editor", this->buildingPalette,
       tipsLabel);
 
-  this->saveAct = new QAction(tr("&Save (As)"), this->mainWindow);
-  this->saveAct->setStatusTip(tr("Save (As)"));
+  this->newAct = new QAction(tr("&New"), this->mainWindow);
+  this->newAct->setStatusTip(tr("New"));
+  this->newAct->setShortcut(tr("Ctrl+N"));
+  this->newAct->setCheckable(false);
+  connect(this->newAct, SIGNAL(triggered()), this, SLOT(New()));
+
+  this->saveAct = new QAction(tr("&Save"), this->mainWindow);
+  this->saveAct->setStatusTip(tr("Save"));
   this->saveAct->setShortcut(tr("Ctrl+S"));
   this->saveAct->setCheckable(false);
   connect(this->saveAct, SIGNAL(triggered()), this, SLOT(Save()));
 
-  this->discardAct = new QAction(tr("&Discard"), this->mainWindow);
-  this->discardAct->setStatusTip(tr("Discard"));
-  this->discardAct->setShortcut(tr("Ctrl+D"));
-  this->discardAct->setCheckable(false);
-  connect(this->discardAct, SIGNAL(triggered()), this, SLOT(Discard()));
-
-  this->doneAct = new QAction(tr("Don&e"), this->mainWindow);
-  this->doneAct->setShortcut(tr("Ctrl+E"));
-  this->doneAct->setStatusTip(tr("Done"));
-  this->doneAct->setCheckable(false);
-  connect(this->doneAct, SIGNAL(triggered()), this, SLOT(Done()));
+  this->saveAsAct = new QAction(tr("&Save As"), this->mainWindow);
+  this->saveAsAct->setStatusTip(tr("Save As"));
+  this->saveAsAct->setShortcut(tr("Ctrl+SHIFT+S"));
+  this->saveAsAct->setCheckable(false);
+  connect(this->saveAsAct, SIGNAL(triggered()), this, SLOT(SaveAs()));
 
   this->exitAct = new QAction(tr("E&xit Building Editor"), this->mainWindow);
   this->exitAct->setStatusTip(tr("Exit Building Editor"));
@@ -87,20 +87,20 @@ BuildingEditor::~BuildingEditor()
 void BuildingEditor::Save()
 {
   gui::editor::Events::saveBuildingEditor(
+    this->buildingPalette->GetModelName());
+}
+
+////////////////////////////////////////////////
+void BuildingEditor::SaveAs()
+{
+  gui::editor::Events::saveAsBuildingEditor(
       this->buildingPalette->GetModelName());
 }
 
 /////////////////////////////////////////////////
-void BuildingEditor::Discard()
+void BuildingEditor::New()
 {
-  gui::editor::Events::discardBuildingEditor();
-}
-
-/////////////////////////////////////////////////
-void BuildingEditor::Done()
-{
-  gui::editor::Events::doneBuildingEditor(
-      this->buildingPalette->GetModelName());
+  gui::editor::Events::newBuildingEditor();
 }
 
 /////////////////////////////////////////////////
@@ -126,9 +126,9 @@ void BuildingEditor::CreateMenus()
   this->menuBar->setSizePolicy(QSizePolicy::Fixed, QSizePolicy::Fixed);
 
   QMenu *fileMenu = this->menuBar->addMenu(tr("&File"));
+  fileMenu->addAction(this->newAct);
   fileMenu->addAction(this->saveAct);
-  fileMenu->addAction(this->discardAct);
-  fileMenu->addAction(this->doneAct);
+  fileMenu->addAction(this->saveAsAct);
   fileMenu->addAction(this->exitAct);
 }